requirements.txt
flask
requests
openai
python-dotenv
pyngrok
psycopg2-binary
gunicorn
